perf(aerogpu): shrink alloc-table SeenIndex scratch to u16
SeenIndex only stores an entry index into tmpEntries, and per-submit entry counts are capped by AEROGPU_KMD_SUBMIT_ALLOCATION_LIST_MAX_COUNT (0xFFFF).

Store SeenIndex as uint16_t instead of UINT to reduce scratch/cache footprint and improve locality.

+++ drivers/aerogpu/kmd/include/aerogpu_kmd.h
@@ -138,7 +138,7 @@ typedef struct _AEROGPU_ALLOC_TABLE_SCRATCH {
 
     struct aerogpu_alloc_entry* TmpEntries;
     uint32_t* Seen;
-    UINT* SeenIndex;
+    uint16_t* SeenIndex;
     uint32_t* SeenEpoch;
     uint32_t Epoch;
